# openapi.yaml
openapi: 3.0.3
info:
  title: Down Dirty 84 AI Tuning MVP API
  version: 0.3.0
servers:
- url: http://localhost:8080
paths:
...
  /api/v1/jobs/{jobId}:
    get:
      summary: Get job
      parameters:
      - name: jobId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
        '404':
          description: Not found
...
  /api/v1/uploads/jobs/{jobId}:
    get:
      summary: List uploads for job
      parameters:
      - name: jobId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
...
components:
...
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: dd84_session